+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.3](https://github.com/at-gmbh/personio-py/tree/v0.2.3) - 2023-02-08
+
+* debug attendance feature
+* add support for attendance paginated API requests
+
 ## [Unreleased](https://github.com/at-gmbh/personio-py/compare/v0.2.2...HEAD)